requirements.txt
python-dotenv~=1.1.1
Flask~=3.1.1
bokeh~=3.7.3
supabase~=2.16.0
pandas~=2.3.0
schedule~=1.2.2
pdfkit~=1.0.0
Jinja2~=3.1.6